+++ protocols/kad/src/protocol.rs
@@ -43,6 +43,9 @@ use wasm_timer::Instant;
 /// The protocol name used for negotiating with multistream-select.
 pub const DEFAULT_PROTO_NAME: &[u8] = b"/ipfs/kad/1.0.0";
 
+/// The default maximum size for a varint length-delimited packet.
+pub const DEFAULT_MAX_PACKET_SIZE: usize = 16 * 1024;
+
 /// Status of our connection to a node reported by the Kademlia protocol.
 #[derive(Copy, Clone, PartialEq, Eq, Debug, Hash)]
 pub enum KadConnectionType {
@@ -167,7 +170,7 @@ impl Default for KademliaProtocolConfig {
     fn default() -> Self {
         KademliaProtocolConfig {
             protocol_name: Cow::Borrowed(DEFAULT_PROTO_NAME),
-            max_packet_size: 4096,
+            max_packet_size: DEFAULT_MAX_PACKET_SIZE,
         }
     }
 }