+++ protocols/kad/CHANGELOG.md
@@ -1,5 +1,8 @@
 # 0.23.0 [unreleased]
 
+- Increase default max packet size from 4KiB to 16KiB.
+  See [issue 1622](https://github.com/libp2p/rust-libp2p/issues/1622).
+
 - Add `Distance::log2` ([PR 1719](https://github.com/libp2p/rust-libp2p/pull/1719)).
 
 - Update `libp2p-swarm` and `libp2p-core`.